requirements.txt
aniso8601          
blinker            
click              
dnspython          
Flask              
Flask-JWT-Extended 
Flask-PyMongo      
Flask-RESTful      
Flask-SQLAlchemy   
greenlet
gunicorn           
itsdangerous       
Jinja2             
MarkupSafe
packaging
pip                
PyJWT              
pymongo
python-dotenv           
pytz              
setuptools         
six    
SQLAlchemy            
typing_extensions 
Werkzeug          
WTForms 